requirements.txt
flask 
sqlalchemy
flask_sqlalchemy
biopython
pandas
python-dotenv
pymysql
requests
beautifulsoup4